+++ src/collective/folderishtypes/at/migration/migrate_plonearticle.py
@@ -12,12 +12,42 @@ import logging
 
 logger = logging.getLogger('collective.folderishtypes PloneArticle migration')
 
+try:
+    from collective.contentleadimage.config import IMAGE_FIELD_NAME
+    from collective.contentleadimage.config import IMAGE_CAPTION_FIELD_NAME
+except ImportError:
+    IMAGE_FIELD_NAME = None
+    IMAGE_CAPTION_FIELD_NAME = None
+
 
 class PloneArticleMigrator(ATItemMigrator):
     src_portal_type = 'PloneArticle'
     src_meta_type = 'PloneArticle'
     dst_portal_type = 'Folderish Document'
     dst_meta_type = 'FolderishDocument'
+
+    def migrate_various(self):
+
+        if IMAGE_FIELD_NAME:
+            # Migrate contentleadimage
+            new_img_field = self.new.getField(IMAGE_FIELD_NAME)
+            if new_img_field:
+                # Else: forgot to configure the type for it? or not supposed to
+                # hold contentlead images
+                img_field = self.old.getField(IMAGE_FIELD_NAME)
+                img = img_field.get(self.old)
+                if img:
+                    new_img_field.set(self.new, img, mimetype=img.content_type)
+
+                    cap_field = self.old.getField(IMAGE_CAPTION_FIELD_NAME)
+                    cap = cap_field.get(self.old)
+                    new_cap_field = self.new.getField(IMAGE_CAPTION_FIELD_NAME)
+                    new_cap_field.set(self.new, cap)
+
+                    logger.info('migrated contentleadimage for %s'
+                                % self.new.absolute_url())
+
+        logger.info('migrated PloneArticle %s' % self.new.absolute_url())
 
     def migrate_files(self):
         """Migrate files from PloneArticle to file contents within the
@@ -44,6 +74,8 @@ class PloneArticleMigrator(ATItemMigrator):
             )
             item_new = self.new.files[id_]
             item_new.setFile(data)
+            item_new.setFilename(item.attachedFile.filename)
+            item_new.setContentType(item.attachedFile.content_type)
 
             item_new.creation_date = DateTime(creation_date)
             item_new.setModificationDate(DateTime(modification_date))
@@ -76,6 +108,8 @@ class PloneArticleMigrator(ATItemMigrator):
             )
             item_new = self.new.images[id_]
             item_new.setImage(data)
+            item_new.setFilename(item.attachedImage.filename)
+            item_new.setContentType(item.attachedImage.content_type)
 
             item_new.creation_date = DateTime(creation_date)
             item_new.setModificationDate(DateTime(modification_date))